Changes timestamps to unix

+++ crates/aptos-telemetry-service/src/clients/mod.rs
@@ -38,7 +38,7 @@ pub mod victoria_metrics_api {
             raw_metrics_body: Bytes,
             extra_labels: Vec<String>,
             encoding: String,
-            timestamp: usize,
+            timestamp: u64,
         ) -> Result<reqwest::Response, anyhow::Error> {
             let mut labels: Vec<(String, String)> = extra_labels
                 .iter()
@@ -47,10 +47,7 @@ pub mod victoria_metrics_api {
             labels.push(("timestamp".into(), timestamp.to_string()));
 
             self.inner
-                .post(format!(
-                    "{}api/v1/import/prometheus",
-                    self.base_url
-                ))
+                .post(format!("{}api/v1/import/prometheus", self.base_url))
                 .bearer_auth(self.auth_token.clone())
                 .header(CONTENT_ENCODING, encoding)
                 .query(&labels)
